Revert "Add additional info popover"
