+++ src/views/main-page-view/components/mobile-timetable/lessons-cards.tsx
@@ -1,6 +1,6 @@
 import type { Timetable } from "@/types/data-access/timetable";
 import { useCallback } from "react";
-import LessonsCard from "./lessons-card";
+import LessonCard from "./lesson-card";
 
 type Props = {
     hours: string[];
@@ -13,24 +13,21 @@ const useLessonsCards = (
     currentDayData: Timetable["data"][number],
     weekParity: "EVEN" | "ODD",
 ) => {
-    const getLessons = useCallback(
+    const getLesson = useCallback(
         (index: number) => {
             const classData =
                 weekParity === "EVEN"
                     ? currentDayData.even
                     : currentDayData.odd;
 
-            // filter and sort (to make sure that the lessons are in the same order)
-            const lessons = classData
-                .filter((l) => l.rowId === index)
-                .sort((a, b) => a.name.localeCompare(b.name));
+            const lesson = classData.find((l) => l.rowId === index) || null;
 
-            return lessons;
+            return lesson;
         },
         [currentDayData, weekParity],
     );
 
-    return { getLessons };
+    return { getLesson };
 };
 
 export default function LessonsCards({
@@ -39,14 +36,14 @@ export default function LessonsCards({
     weekDayIndex,
     weekParity,
 }: Props) {
-    const { getLessons } = useLessonsCards(currentDayData, weekParity);
+    const { getLesson } = useLessonsCards(currentDayData, weekParity);
 
     return (
         <div className="xxs:mt-6 *:border-border-muted *:first:border-t-border-muted mt-4 flex flex-col px-2 pb-10 *:border-b-1 *:first:border-t-1">
             {hours.map((hour, index) => (
-                <LessonsCard
+                <LessonCard
                     key={`timetable-lesson-${index}`}
-                    lessons={getLessons(index)}
+                    lesson={getLesson(index)}
                     hour={hour}
                     weekDayIndex={weekDayIndex}
                 />
